+++ migrations/m220129_085202_create_rename_plm_panned_some_table.php
@@ -14,23 +14,11 @@ class m220129_085202_create_rename_plm_panned_some_table extends Migration
     {
         $this->renameTable('plm_planned_stop', 'plm_stops');
         $this->addColumn('{{%plm_stops}}','stopping_type',$this->smallInteger());
-        // drops foreign key for table `{{%plm_documents}}`
-        $this->dropForeignKey(
-            '{{%fk-plm_unplanned_stop-doc_id}}',
-            '{{%plm_unplanned_stop}}'
-        );
 
-        // drops index for column `doc_id`
-        $this->dropIndex(
-            '{{%idx-plm_unplanned_stop-doc_id}}',
-            '{{%plm_unplanned_stop}}'
-        );
-        $this->dropIndex(
-            '{{%idx-plm_unplanned_stop-status_id}}',
-            '{{%plm_unplanned_stop}}'
-        );
-        $this->dropTable('{{%plm_unplanned_stop}}');
-
+        $q = $this->db->createCommand("SELECT EXISTS (SELECT FROM pg_tables WHERE schemaname = 'public' AND tablename  = 'plm_unplanned_stop' );")->queryScalar();
+        if ($q) {
+            $this->dropTable('{{%plm_unplanned_stop}}');
+        }
     }
     /**
      * {@inheritdoc}
@@ -38,5 +26,6 @@ class m220129_085202_create_rename_plm_panned_some_table extends Migration
     public function safeDown()
     {
         $this->renameTable('plm_stops','plm_planned_stop');
+        $this->dropColumn('{{%plm_planned_stop}}', 'stopping_type');
     }
 }
